% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{angleD}
\alias{angleD}
\title{Calcul de l'angle D  entre  -90 <=  angD  < +270  retourne en radian}
\usage{
angleD(X, Y)
}
\description{
Calcul de l'angle D  entre  -90 <=  angD  < +270  retourne en radian
}
