% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.eigen.tensors.all}
\alias{anisotropie.eigen.tensors.all}
\title{Calcul des tenseurs d'anisotropie pour tous les spécimens d'une liste de mesures}
\usage{
anisotropie.eigen.tensors.all(Data.mesures, Data.info, etape.value)
}
\value{
un data.frame avec les colonnes "L1", "L1.Inc", "L1.Dec", "L2", "L2.Inc", "L2.Dec", "L3", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul des tenseurs d'anisotropie pour tous les spécimens d'une liste de mesures
}
