% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.mean.matrix}
\alias{anisotropie.mean.matrix}
\title{Calcul de la matrice moyenne symétrisée d'anisotropie moyen}
\usage{
anisotropie.mean.matrix(Data.mesures, Data.number, step.value,
  step.code = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB"))
}
\arguments{
\item{Data.mesures}{data.frame contenant les mesures}

\item{Data.number}{numéro des échantillons}

\item{step.value}{typiquement la température des mesures d'anisotropie}
}
\value{
une matrice carrée 3 x3
}
\description{
Calcul de la matrice moyenne symétrisée d'anisotropie moyen
}
\seealso{
\code{\link{anisotropie.mean.eigen.tensor}}
}
