% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie}
\alias{anisotropie}
\title{Calcul du tenseur et de la matrice d'anisotropie}
\usage{
anisotropie(mesures, etape.value, etape.sigle = c("Z+", "Z-", "X+", "X-",
  "Y+", "Y-", "ZB"), volume = 1, TH = 1, ...)
}
\description{
Calcul du tenseur et de la matrice d'anisotropie
}
