% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.tensors.names}
\alias{anisotropie.tensors.names}
\title{Calcul des tenseurs d'anisotropie pour une liste de nom de spécimen}
\usage{
anisotropie.tensors.names(names, Data.mesures, Data.info, etape.value)
}
\description{
Calcul des tenseurs d'anisotropie pour une liste de nom de spécimen
}
