% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.tensors.numbers}
\alias{anisotropie.tensors.numbers}
\title{Calcul des tenseurs d'anisotropie pour une liste de numéro de spécimen}
\usage{
anisotropie.tensors.numbers(numbers, Data.mesures, Data.info, etape.value)
}
\description{
Calcul des tenseurs d'anisotropie pour une liste de numéro de spécimen
}
