% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.eigen}
\alias{anisotropy.eigen}
\title{Calcul du vecteur propre et de la matrice d'anisotropie pour un spécimen}
\usage{
anisotropy.eigen(
  mesures,
  step.value,
  step.code = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB"),
  volume = 1,
  TH = 1,
  ...
)
}
\arguments{
\item{mesures}{data.frame contenant les mesures}

\item{step.value}{typiquement la température des mesures d'anisotropie}

\item{step.code}{code indiquant les étapes. les noms peuvent changer, mais pas l'ordre}

\item{volume}{la valeur du volume du spécimen}

\item{TH}{la valeur du champ appliqué}
}
\value{
un data.frame avec les colonnes "L1", "L1.Inc", "L1.Dec", "L2", "L2.Inc", "L2.Dec", "L3", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul du vecteur propre et de la matrice d'anisotropie pour un spécimen
}
