% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.eigen.tensor.total}
\alias{anisotropy.eigen.tensor.total}
\title{Calcul des directions et valeurs des vecteurs propres d'anisotropie totale pour un spécimen à une certaine température}
\usage{
anisotropy.eigen.tensor.total(
  mesures,
  step.value,
  step.code = c("ZZ", "XX", "YY", "ZB")
)
}
\arguments{
\item{mesures}{data.frame contenant les mesures}

\item{step.value}{typiquement la température des mesures d'anisotropie}

\item{step.code}{code indiquant les étapes. Les noms peuvent changer, mais pas l'ordre}
}
\value{
un data.frame avec les colonnes "L1", "L1.Inc", "L1.Dec", "L2", "L2.Inc", "L2.Dec", "L3", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul des directions et valeurs des vecteurs propres d'anisotropie totale pour un spécimen à une certaine température
}
\seealso{
\code{\link{anisotropy.eigen.matrix}}
}
