% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.mean.eigen.tensor}
\alias{anisotropy.mean.eigen.tensor}
\title{Calcul de la direction et des valeurs du vecteur propre d'anisotropie moyen}
\usage{
anisotropy.mean.eigen.tensor(
  Data.mesures,
  Data.number,
  step.value,
  step.code = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB")
)
}
\arguments{
\item{Data.mesures}{data.frame contenant les mesures}

\item{Data.number}{numéro des échantillons}

\item{step.value}{typiquement la température des mesures d'anisotropie}
}
\value{
un data.frame avec les colonnes "L1", "L2", "L3", "L1.Inc", "L1.Dec", "L2.Inc", "L2.Dec", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul de la direction et des valeurs du vecteur propre d'anisotropie moyen
}
\seealso{
\code{\link{anisotropy.eigen.matrix}}, \code{\link{anisotropy.mean.matrix}}
}
