% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.mean.matrix.multi}
\alias{anisotropy.mean.matrix.multi}
\title{Calcul la matrice d'anisotropie moyenne symmétrisée pour 2 valeurs de températures différentes
step.value est un vecteur des 2 températures dans l'ordre des deux data frames}
\usage{
anisotropy.mean.matrix.multi(
  dat1,
  dat2,
  step.value,
  step.code1 = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB"),
  step.code2 = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB")
)
}
\arguments{
\item{dat1}{mesures d'anisotropie à la température T1}

\item{dat2}{mesures d'anisotropie à la température T2}

\item{step.value}{vecteur des températures T1 et T2}

\item{matrix}{matrice 3 x 3 correspondant à la matrice (symétrique et normalisé) à utiliser pour la correction}
}
\value{
une matrice
}
\description{
Calcul la matrice d'anisotropie moyenne symmétrisée pour 2 valeurs de températures différentes
step.value est un vecteur des 2 températures dans l'ordre des deux data frames
}
