% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.mean.matrix.total}
\alias{anisotropy.mean.matrix.total}
\title{Calcul de la matrice moyenne symétrisée d'anisotropie moyen (anisotropie totale)}
\usage{
anisotropy.mean.matrix.total(
  Data.mesures,
  Data.number,
  step.value,
  step.code = c("ZZ", "XX", "YY", "ZB")
)
}
\arguments{
\item{Data.mesures}{data.frame contenant les mesures}

\item{Data.number}{numéro des échantillons}

\item{step.value}{typiquement la température des mesures d'anisotropie}
}
\value{
une matrice carrée 3 x3
}
\description{
Calcul de la matrice moyenne symétrisée d'anisotropie moyen (anisotropie totale)
}
\seealso{
\code{\link{anisotropy.mean.eigen.tensor}}
}
