% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropy.tensor}
\alias{anisotropy.tensor}
\title{Calcul de la direction et des valeurs du vecteur propre d'anisotropie moyen à partir de la matrice
d'anisotropie et non  pas du data.frame de mesures
Il faut une matrice 3 x3}
\usage{
anisotropy.tensor(matrix)
}
\arguments{
\item{matrix}{matrice 3 x 3 correspondant à la matrice (symétrique et normalisé) à utiliser pour la correction}
}
\value{
un data.frame avec les colonnes "L1", "L2", "L3", "L1.Inc", "L1.Dec", "L2.Inc", "L2.Dec", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul de la direction et des valeurs du vecteur propre d'anisotropie moyen à partir de la matrice
d'anisotropie et non  pas du data.frame de mesures
Il faut une matrice 3 x3
}
