% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{arai}
\alias{arai}
\title{Standart calcul and plot to study paleo and archeo magnetic magnétisation
  by convention we use the two last character of the variable step to indicate the type of manip
  example 100RA , 100 is the temperature (step.value) , R is the sens of the magnetisation and A is the name of the step (step.name)}
\usage{
arai(
  mesures,
  relative = TRUE,
  verbose = TRUE,
  show.plot = TRUE,
  TH = 60,
  aim.coef = 1e-10 * 1e+06,
  step.J0 = "20N0",
  show.step.value = FALSE,
  R.mark = "R",
  V.mark = "V",
  P.mark = "P",
  L.mark = "L",
  Q.mark = "Q",
  pt.col = "blue",
  loop.col = "forestgreen",
  begin.step.value = 0,
  end.step.value = 1000
)
}
\arguments{
\item{mesures}{data.frame with the package convention format}

\item{relative}{plot with a relative value in percent}

\item{verbose}{show comment}

\item{show.plot}{display the plot}

\item{TH}{lab field}

\item{aim.coef}{used to correct the measurement often 1E-10x1E6 x volume}

\item{step.J0}{is the step of the natural magnetisation eg "0N0", default 20N0}

\item{show.step.value}{display on the plot the value of each step}

\item{R.mark}{= 'R', V.mark = 'V' conventional notation to mark the sens of the magnetisation with the step name}

\item{P.mark}{= 'P' mark the pTRM check loop}

\item{L.mark}{= "L", Q.mark = "Q" mark to indiquate the slow cooling step and the quick (fast-normal) cooling step}

\item{pt.col}{the color of the line and the plot}

\item{loop.col}{the color of the line showing the the check loop process}

\item{begin.step.value}{the first step.value (temperature) used to determinate the magnetic field}

\item{end.step.value}{the last step.value (temperature) used to determinate the magnetic field}
}
\description{
Standart calcul and plot to study paleo and archeo magnetic magnétisation
  by convention we use the two last character of the variable step to indicate the type of manip
  example 100RA , 100 is the temperature (step.value) , R is the sens of the magnetisation and A is the name of the step (step.name)
}
\references{
Coe 1978 : DOI: 10.1029/JB083iB04p01740
Prévost et Al. 1985 DOI: 10.1029/JB090iB12p10417
}
