% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{calcul.vecteur.cartesien.Z}
\alias{calcul.vecteur.cartesien.Z}
\title{Valeur de Z pour I et D}
\usage{
calcul.vecteur.cartesien.Z(inc, dec, aim = 1)
}
\arguments{
\item{inc}{liste des inclinaisons en radian}

\item{des}{liste des déclinaisons en radian}
}
\description{
Valeur de Z pour I et D
}
