% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{calcul.vecteur.polaire.D}
\alias{calcul.vecteur.polaire.D}
\title{Déclinaison pour X, Y et Z
#' @return angle en degré}
\usage{
calcul.vecteur.polaire.D(X, Y, Z)
}
\description{
Déclinaison pour X, Y et Z
#' @return angle en degré
}
