% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{calcul.vecteur.polaire}
\alias{calcul.vecteur.polaire}
\title{Coordonnées polaires pour X, Y et Z}
\usage{
calcul.vecteur.polaire(X, Y, Z)
}
\value{
angle en degré
}
\description{
Coordonnées polaires pour X, Y et Z
}
\seealso{
polaire
}
