% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{cartesien}
\alias{cartesien}
\title{Valeur de X, Y et Z pour I et D}
\usage{
cartesien(inc, dec, aim = 1)
}
\arguments{
\item{inc}{liste des inclinaisons en degrés}

\item{des}{liste des déclinaisons en degrés}
}
\description{
Valeur de X, Y et Z pour I et D
}
\seealso{
\code{\link{calcul.vecteur.cartesien}} , \code{\link{polaire}}
}
