% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{composante.partielle}
\alias{composante.partielle}
\title{Calcul le vecteur de la composante partielle
et retourne aussi le MAD}
\usage{
composante.partielle(TabX, TabY, TabZ, en0 = FALSE)
}
\arguments{
\item{en0}{permet de calculer la composante qui passe par l'origine (0, 0)}
}
\value{
une data.frame "X", "Y", "Z", "I", "D", "F", "Sl", "MAD", "DANG"
}
\description{
Calcul le vecteur de la composante partielle
et retourne aussi le MAD
}
