% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{composante.partielle.T1T2}
\alias{composante.partielle.T1T2}
\title{Calcul le vecteur de la composante partielle en calculant la composante entre les étapes T1 et T2
par défaut la fonction n'affiche pas les étapes d'anisotropie}
\usage{
composante.partielle.T1T2(Data, T1 = NULL, T2 = NULL,
  corr.ani = FALSE, ani.step.value = NULL, step.code = c("Z+", "Z-",
  "X+", "X-", "Y+", "Y-", "ZB"), en0 = FALSE)
}
\arguments{
\item{corr.ani}{corrige de l'anisotropie}

\item{en0}{permet de calculer la composante qui passe par l'origine (0, 0)}
}
\value{
une data.frame "X", "Y", "Z", "I", "D", "F", "Sl", "MAD", "DANG"
}
\description{
Calcul le vecteur de la composante partielle en calculant la composante entre les étapes T1 et T2
par défaut la fonction n'affiche pas les étapes d'anisotropie
}
\seealso{
composante.partielle, zijderveld1.T1T2
}
