% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{cooling.rate}
\alias{cooling.rate}
\title{calculus to determinate the cooling rates and evolution factors to apply the cooling rate correction
  by convention we use the two last characters of the different steps to indicate the type of experimentation
  example 100RA , 100 is the temperature (step.value) , R is the direction of the magnetization and A is the name of the step (step.name)}
\usage{
cooling.rate(
  mesures,
  verbose = TRUE,
  TH = 60,
  aim.coef = 1/(10.8 * 1e-06),
  R.mark = "R",
  V.mark = "V",
  L.mark = "L",
  Q.mark = "Q",
  begin.step.value = 0,
  end.step.value = 1000,
  digits = 2,
  Cooling.stepvalue = NA
)
}
\arguments{
\item{mesures}{data.frame with the package convention format}

\item{verbose}{show comment}

\item{TH}{lab field}

\item{aim.coef}{used to correct the measurement often 1E-10x1E6 x volume}

\item{R.mark}{= 'R' points the direction of the magnetization with the step name}

\item{V.mark}{= 'V' points the direction of the magnetization with the step name}

\item{L.mark}{= "L" points the slow cooling step and}

\item{Q.mark}{= "Q" points the quick (fast-normal) cooling step}

\item{begin.step.value}{the first step.value (temperature) used to determinate the magnetic field}

\item{end.step.value}{the last step.value (temperature) used to determinate the magnetic field}

\item{Cooling.stepvalue}{temperature of experimentation}
}
\description{
calculus to determinate the cooling rates and evolution factors to apply the cooling rate correction
  by convention we use the two last characters of the different steps to indicate the type of experimentation
  example 100RA , 100 is the temperature (step.value) , R is the direction of the magnetization and A is the name of the step (step.name)
}
\references{
Coe 1978 : DOI: 10.1029/JB083iB04p01740
Prévost et Al. 1985 DOI: 10.1029/JB090iB12p10417
}
