% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{correction.archeo}
\alias{correction.archeo}
\title{Correction de repère d'orientement avec la norme archeomagnetique
Pour les Matériaux déplacés:
on utilise Inc et Az pour faire la correction de #, c'est à dire repère de carrotage perpendiculaire au bord long
cela revient à tourner dans le sens trigonométrique dans les stéréogrammes}
\usage{
correction.archeo(mes, inc = 0, az = 0)
}
\description{
Correction de repère d'orientement avec la norme archeomagnetique
Pour les Matériaux déplacés:
on utilise Inc et Az pour faire la correction de #, c'est à dire repère de carrotage perpendiculaire au bord long
cela revient à tourner dans le sens trigonométrique dans les stéréogrammes
}
\seealso{
\code{\link{correction.paleo}}, \code{\link{correction.bevel}}, \code{\link{correction.bending}}
}
