% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{correction.bending}
\alias{correction.bending}
\title{Correction de pli = bending
Note: la correction de pli- bending se fait aprés la correction de repère}
\usage{
correction.bending(mes, dip, str)
}
\description{
Correction de pli = bending
Note: la correction de pli- bending se fait aprés la correction de repère
}
\seealso{
\code{\link{correction.paleo}}, \code{\link{correction.archeo}}, \code{\link{correction.bevel}}
}
