% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{correction.bevel}
\alias{correction.bevel}
\title{Correction biseau - bevel
En archéomagnétisme correspond à une erreur de carottage par rapport à l'axe théorique de carottage,
à une mauvaise perpendicularité du carottage généralement}
\usage{
correction.bevel(mes, theta = 0, psi = 0)
}
\arguments{
\item{psi}{est compté positif dans le sens des aiguilles d'une montre et donc négatif dans le sens inverse...
Note: la correction de biseau - bevel se fait avant la correction de repère}

\item{Theta}{est exprimé entre 0 et 90° (toujours positif)}
}
\description{
Correction biseau - bevel
En archéomagnétisme correspond à une erreur de carottage par rapport à l'axe théorique de carottage,
à une mauvaise perpendicularité du carottage généralement
}
\seealso{
\code{\link{correction.paleo}}, \code{\link{correction.archeo}}, \code{\link{correction.bending}}
}
