% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{correction.paleo}
\alias{correction.paleo}
\title{C'est la correction de repère utilisée en paléomag}
\usage{
correction.paleo(mes, inc = 0, az = 0)
}
\arguments{
\item{Inc}{=0   si carotte verticale et   Inc=90° si carotte horizontale}

\item{Az}{est compté positif dans le sens des aiguilles d'une montre}
}
\description{
C'est la correction de repère utilisée en paléomag
}
\seealso{
\code{\link{correction.archeo}}, \code{\link{correction.bevel}} , \code{\link{correction.bending}}
}
