% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{D.AM}
\alias{D.AM}
\title{Convertie la déclinaison dans la convention AM}
\usage{
D.AM(d)
}
\arguments{
\item{d}{déclinaison en degré}
}
\value{
angle en degré en -90° et 270°
}
\description{
Convertie la déclinaison dans la convention AM
}
