% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{D.pal}
\alias{D.pal}
\title{Convertie la déclinaison dans la convention paleomag}
\usage{
D.pal(d)
}
\arguments{
\item{d}{déclinaison en degrés}
}
\value{
déclinaison entre 0 et 360°
}
\description{
Convertie la déclinaison dans la convention paleomag
}
