% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{DD.to.DMS}
\alias{DD.to.DMS}
\title{Conversion de Degré Décimal en Degres minute second}
\usage{
DD.to.DMS(degre)
}
\arguments{
\item{degre}{degrés décimal}
}
\description{
Conversion de Degré Décimal en Degres minute second
}
\seealso{
DMS.to.DD
}
