% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{desaim}
\alias{desaim}
\title{Tracer d'un diagramme de désaimantation}
\usage{
desaim(Data, F = NULL, point.col = "blue3", pch = 21, type = "b",
  xlab = "°C", ylab = "", main = NULL, names = NA,
  normalize = TRUE, step.J0 = NULL, new = TRUE, ...)
}
\arguments{
\item{normalize}{permet de comparer l'évolution de l'aimanation quelque soit l'amplitude en visualisant le résultat comme un pourcentage du maximum}
}
\description{
Tracer d'un diagramme de désaimantation
}
