% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{DMS.to.DD}
\alias{DMS.to.DD}
\title{Conversion de Degres minute second en Degré Décimal}
\usage{
DMS.to.DD(degre, minute, second)
}
\arguments{
\item{degre}{degrés entier}

\item{minute}{minute entière}

\item{second}{seconde en décimal}
}
\description{
Conversion de Degres minute second en Degré Décimal
}
\seealso{
DD.to.DMS
}
