% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{extract.mesures.specimen.name}
\alias{extract.mesures.specimen.name}
\title{Extraction des mesures correspondant à un nom}
\usage{
extract.mesures.specimen.name(specimen.name, list.mesure)
}
\description{
Extraction des mesures correspondant à un nom
}
