% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{extract.mesures.specimen.number}
\alias{extract.mesures.specimen.number}
\title{Extraction des mesures correspondant à un numéro}
\usage{
extract.mesures.specimen.number(specimen.number, list.mesure)
}
\description{
Extraction des mesures correspondant à un numéro
}
