% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{genere.AMD}
\alias{genere.AMD}
\title{Fonction de création de fichier pour les magnétomètres, pour des matériaux déplacés ----}
\usage{
genere.AMD(
  file.AMD = "fichier.AMD",
  list.ech,
  shape = "Cyl",
  encoding = "macroman"
)
}
\arguments{
\item{encoding}{mettre "macroman" pour les mesures au molspin}
}
\description{
Fonction de création de fichier pour les magnétomètres, pour des matériaux déplacés ----
}
