% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{genere.AMP}
\alias{genere.AMP}
\title{Fonction de création de fichier pour les magnétomètres, pour des structures en place ----}
\usage{
genere.AMP(
  file.AMP = "fichier.AMP",
  list.ech,
  shape = "Cyl",
  encoding = "macroman"
)
}
\arguments{
\item{encoding}{mettre "macroman" pour les mesures au molspin}
}
\description{
Fonction de création de fichier pour les magnétomètres, pour des structures en place ----
}
