% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{igrf13syn}
\alias{igrf13syn}
\title{This is a synthesis routine for the 13th generation IGRF as agreed
   in December 2019 by IAGA Working Group V-MOD. It is valid 1900.0 to
    2025.0 inclusive. Values for dates from 1945.0 to 2015.0 inclusive are
    definitive, otherwise they are non-definitive.
  INPUT}
\usage{
igrf13syn(isv, date, itype, alt, colat, elong)
}
\arguments{
\item{isv}{= 0 if main-field values are required,  = 1 if secular variation values are required}

\item{date}{= year A.D. Must be greater than or equal to 1900.0 and less than or equal to 2030.0. Warning message is given
for dates greater than 2025.0. Must be double precision.}

\item{itype}{= 1 if geodetic (spheroid);  = 2 if geocentric (sphere)}

\item{alt}{= height in km above sea level if itype = 1;   = distance from centre of Earth in km if itype = 2 (>3485 km)}

\item{colat}{= colatitude (0-180)}

\item{elong}{= east-longitude (0-360)
    alt, colat and elong must be double precision.
  OUTPUT
    x     = north component (nT) if isv = 0, nT/year if isv = 1
    y     = east component (nT) if isv = 0, nT/year if isv = 1
    z     = vertical component (nT) if isv = 0, nT/year if isv = 1
    f     = total intensity (nT) if isv = 0, rubbish if isv = 1

    To get the other geomagnetic elements (D, I, H and secular
    variations dD, dH, dI and dF) use routines ptoc and ptocsv.

    Adapted from 8th generation version to include new maximum degree for
    main-field models for 2000.0 and onwards and use WGS84 spheroid instead
    of International Astronomical Union 1966 spheroid as recommended by IAGA
    in July 2003. Reference radius remains as 6371.2 km - it is NOT the mean
    radius (= 6371.0 km) but 6371.2 km is what is used in determining the
    coefficients. Adaptation by Susan Macmillan, August 2003 (for
    9th generation), December 2004, December 2009 & December 2014;
    by William Brown, December 2019, February 2020.

    Coefficients at 1995.0 incorrectly rounded (rounded up instead of
    to even) included as these are the coefficients published in Excel
    spreadsheet July 2005.
igrf13syn(isv=isv, date=date, itype=itype , alt=alt, colat= colat, elong=elong)}
}
\description{
This is a synthesis routine for the 13th generation IGRF as agreed
   in December 2019 by IAGA Working Group V-MOD. It is valid 1900.0 to
    2025.0 inclusive. Values for dates from 1945.0 to 2015.0 inclusive are
    definitive, otherwise they are non-definitive.
  INPUT
}
\examples{
isv <- 1
date <- 2000
itype <- 1
alt <- 50
colat <- 90 - 45
elong <- 10

}
\references{
Thébault, E., Finlay, C.C., Beggan, C.D. et al. International Geomagnetic Reference Field: the 12th generation. Earth Planet Sp 67, 79 (2015).
https://doi.org/10.1186/s40623-015-0228-9 (\link{https://earth-planets-space.springeropen.com/articles/10.1186/s40623-015-0228-9}): the 12th generation, Erwan Thébault, Christopher C Finlay, Ciarán D Beggan, Patrick Alken, Julien Aubert, Olivier Barrois, Francois Bertrand, Tatiana Bondar, Axel Boness, Laura Brocco, Elisabeth Canet, Aude Chambodut, Arnaud Chulliat, Pierdavide Coïsson, Francois Civet, Aimin Du, Alexandre Fournier, Isabelle Fratter, Nicolas Gillet, Brian Hamilton, Mohamed Hamoudi, Gauthier Hulot, Thomas Jager, Monika Korte, Weijia Kuang, Xavier Lalanne, Benoit Langlais, Jean-Michel Léger, Vincent Lesur, Frank J Lowes et al. Earth, Planets and Space 2015, 67:79 (27 May 2015)
}
\seealso{
\code{\link{https://www.ngdc.noaa.gov/IAGA/vmod/igrf.html}}
}
