% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{lambert.ID.circle}
\alias{lambert.ID.circle}
\title{Trace un cercle sur le diagramme lambert}
\usage{
lambert.ID.circle(
  i.mean,
  d.mean,
  delta,
  inc.lim = NULL,
  dec.min = -90,
  dec.max = 270,
  col = par("fg"),
  clockwise = TRUE,
  absolue = TRUE,
  lty = par("lty"),
  ...
)
}
\arguments{
\item{col}{définie la couleur de la ligne}

\item{absolue}{permet de tracer la partie inclinaison négative des cercles}
}
\description{
Trace un cercle sur le diagramme lambert
}
