% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{lambert.ID}
\alias{lambert.ID}
\title{Place des points I et D dans un repère Lambert}
\usage{
lambert.ID(inc, dec, pt.names = NA, labels = NA, label.pos = NULL,
  radlab = FALSE, start = 0, clockwise = TRUE, label.prop = 1.1,
  main = "", xlab = "", ylab = "", line.col = par("fg"),
  lty = par("lty"), lwd = par("lwd"), mar = c(2, 2, 3, 2),
  show.grid = TRUE, show.grid.labels = 10, show.radial.grid = TRUE,
  grid.col = "gray", grid.bg = "transparent", grid.left = FALSE,
  grid.unit = NULL, point.symbols = 1, point.col = par("fg"),
  bg = point.col, inc.lim = NULL, radial.labels = NULL,
  boxed.radial = TRUE, poly.col = NA, dec.min = -90, dec.max = 270,
  new = TRUE, pch = 21, ...)
}
\description{
Place des points I et D dans un repère Lambert
}
