% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{lambert.ID.specimen}
\alias{lambert.ID.specimen}
\title{Place des points I et D dans un repère Lambert avec un data.frame}
\usage{
lambert.ID.specimen(
  Data,
  D,
  pt.names = "",
  labels = NA,
  label.pos = NULL,
  radlab = FALSE,
  start = 0,
  clockwise = TRUE,
  label.prop = 1.1,
  main = NULL,
  xlab = "",
  ylab = "",
  line.col = "blue3",
  lty = par("lty"),
  lwd = par("lwd"),
  mar = c(2, 2, 3, 2),
  show.grid = TRUE,
  show.grid.labels = 10,
  show.radial.grid = TRUE,
  grid.col = "lightgray",
  grid.bg = "transparent",
  grid.left = FALSE,
  grid.unit = NULL,
  type = "o",
  pt.col = "blue3",
  bg = pt.col,
  inc.lim = c(0, 90),
  radial.labels = NULL,
  boxed.radial = TRUE,
  poly.col = NA,
  add = FALSE,
  dec.min = -90,
  dec.max = 270,
  new = TRUE,
  pch = 21,
  ...
)
}
\description{
Place des points I et D dans un repère Lambert avec un data.frame
}
