% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{lambert}
\alias{lambert}
\title{Place des points I et D dans un repère Lambert avec un data.frame}
\usage{
lambert(data, pt.names = NULL, labels = NA, label.pos = NULL,
  radlab = FALSE, start = 0, clockwise = TRUE, label.prop = 1.1,
  main = "", xlab = "", ylab = "", line.col = par("fg"),
  lty = par("lty"), lwd = par("lwd"), mar = c(2, 2, 3, 2),
  show.grid = TRUE, show.grid.labels = 10, show.radial.grid = TRUE,
  grid.col = "gray", grid.bg = "transparent", grid.left = FALSE,
  grid.unit = NULL, point.symbols = 1, point.col = par("fg"),
  bg = point.col, inc.lim = NULL, radial.labels = NULL,
  boxed.radial = TRUE, poly.col = NA, dec.min = -90, dec.max = 270,
  new = TRUE, pch = 21, ...)
}
\arguments{
\item{data}{data.frame avec les variables $I d'inclinaison et $D de déclinaison}

\item{pt.names}{Correspond à la liste des noms des points. Laissée vide n'affiche rien. Si on met pt.names = "", cela affiche les noms}

\item{label.pos}{Séquence de valeur à afficher en I et D, voir fonction lambert.ID.grid}

\item{show.grid}{permet d'afficher une grille en toile d'araigné sur le fond. Mettre à FALSE, si on superpose des diagrammes}

\item{show.grid.labels}{permet de changer échelle des graduations}

\item{point.symbols}{défini la forme de points, correspond exactement au pch de la fonction points()}

\item{inc.lim}{permet de restreindre l'affichage sur une étendue d'inclinaison. Ex: inc.lim = c(45, 90). Laissée à NULL, le diagramme s'addapte aux données}

\item{dec.max}{permet de restreindre l'étendue en déclinaison}

\item{new}{permet d'initialiser la sortie graphique. Mettre à FALSE, si on superpose des diagrammes.}

\item{pch}{: permet de changer la forme du symbole}
}
\description{
Place des points I et D dans un repère Lambert avec un data.frame
}
