% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{n_arcsin}
\alias{n_arcsin}
\title{Calcul l'arc sinus d'un réel renvoie un angle entre -PI/2 et +PI/2 ----}
\usage{
n_arcsin(x)
}
\description{
Calcul l'arc sinus d'un réel renvoie un angle entre -PI/2 et +PI/2 ----
}
