% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.flin}
\alias{plot.flin}
\title{Tracer d'un diagramme de Flin}
\usage{
\method{plot}{flin}(Data, Data.F12 = NULL, pt.names = NULL,
  point.col = "blue3", pch = 21, type = "p", xlab = "F23",
  ylab = "F12", main = "Flin diagram", absolue = TRUE, new = TRUE)
}
\description{
Tracer d'un diagramme de Flin
}
