% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.circle}
\alias{plot.lambert.ID.circle}
\title{Trace un cercle sur le diagramme lambert}
\usage{
\method{plot}{lambert.ID.circle}(i.mean, d.mean, delta, inc.lim = NULL,
  dec.min = -90, dec.max = 180, col = par("fg"), clockwise = TRUE,
  absolue = TRUE, ...)
}
\description{
Trace un cercle sur le diagramme lambert
}
