% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.field}
\alias{plot.lambert.ID.field}
\title{Trace un cercle avec les champs de l'inclinaison possible}
\usage{
\method{plot}{lambert.ID.field}(data, dec = NULL, pt.names = NA,
  field = c(50, 75), inc.lim = c(0, 90), dec.min = -90,
  dec.max = 270, col = par("fg"), main = "", point.col = par("fg"),
  bg = point.col, label.pos = NULL, show.grid = TRUE, new = TRUE)
}
\arguments{
\item{-}{field : définit la zone possible du champ}
}
\description{
Trace un cercle avec les champs de l'inclinaison possible
}
