% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.grid}
\alias{plot.lambert.ID.grid}
\title{plot.lambert.ID.grid}
\usage{
\method{plot}{lambert.ID.grid}(main = "", xlab = "", ylab = "",
  labels = NA, label.pos = NULL, radlab = FALSE, start = 0,
  clockwise = FALSE, label.prop = 1.1, grid.col = "gray",
  grid.bg = "transparent", show.radial.grid = TRUE,
  labels.precision = 0, dec.min = -90, dec.max = 270, inc.min = 0,
  inc.max = 90, new = TRUE, ...)
}
\arguments{
\item{label.pos}{séquence de valeur à afficher en I et D, attention format particulier !!
# Pour choisir les graduations
label.pos = NULL
label.pos$I = seq(0, 90, by=20)
label.pos$D = seq(0, 90, by=10)
et pour paleomag : lab.pos$D = c(seq(270, 350, by=10), seq(0, 90, by=10))
Label.pos : doit être dans l'étendu des dex.min, dec.max, inc.min, inc.max}

\item{radlab}{écrit les labels sous forme d'étoile}
}
\description{
plot.lambert.ID.grid
}
