% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.point}
\alias{plot.lambert.ID.point}
\title{Place des points I et D dans un repère Lambert}
\usage{
\method{plot}{lambert.ID.point}(inc, dec, pt.names = NA, labels = NA,
  label.pos = NULL, start = 0, clockwise = TRUE, lty = par("lty"),
  mar = c(2, 2, 3, 2), point.col = par("fg"), bg = point.col,
  inc.lim = NULL, dec.min = -90, dec.max = 270, new = TRUE,
  pch = 21, ...)
}
\description{
Place des points I et D dans un repère Lambert
}
