% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.position}
\alias{plot.lambert.ID.position}
\title{Place des points I et D dans un repère Lambert et dessine le symbole}
\usage{
\method{plot}{lambert.ID.position}(data, declinaisons = NULL,
  position = "P", pt.names = NA, labels = NA, label.pos = NULL,
  radlab = FALSE, start = 0, clockwise = TRUE, label.prop = 1.1,
  main = "Position auto", xlab = "", ylab = "",
  line.col = par("fg"), lty = par("lty"), lwd = par("lwd"),
  mar = c(2, 2, 3, 2), show.grid = TRUE, show.grid.labels = 10,
  show.radial.grid = TRUE, grid.col = "gray",
  grid.bg = "transparent", grid.left = FALSE, grid.unit = NULL,
  point.symbols = 1, point.col = par("fg"), bg = point.col,
  inc.lim = c(0, 90), radial.labels = NULL, boxed.radial = TRUE,
  poly.col = NA, add = FALSE, dec.min = -90, dec.max = 90,
  new = TRUE, ...)
}
\description{
Place des points I et D dans un repère Lambert et dessine le symbole
}
