% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.lambert.ID.tensors}
\alias{plot.lambert.ID.tensors}
\title{Tracer dans le diagramme Lambert/ Stéréo d'un tenseur}
\usage{
\method{plot}{lambert.ID.tensors}(Data, point.col = "blue3",
  new = TRUE, ...)
}
\description{
Tracer dans le diagramme Lambert/ Stéréo d'un tenseur
}
