% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.zjiderveld1}
\alias{plot.zjiderveld1}
\title{Trace un diagramme de Zijderveld type 1}
\usage{
\method{plot}{zjiderveld1}(Data, Y = NULL, Z = NULL, pt.names = "",
  main = NULL, panel.first = NULL, pt.col = c("forestgreen",
  "blue3"), isometric = TRUE, ylim = NULL, legend.pos = NULL,
  legend.txt = c("(Y, X)", "(Y, Z)"), new = TRUE, ...)
}
\arguments{
\item{panel.first}{= grid() : affiche une grille}

\item{legend.pos}{= NULL : n'affiche pas la legende. legend.pos = "topleft" affiche en haut à gauche}

\item{pt.name}{= "": n'affiche rien. pt.name = NULL: affiche les étapes si Data est une data.frame}
}
\description{
Trace un diagramme de Zijderveld type 1
}
