% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.AM.mesures}
\alias{read.AM.mesures}
\title{Lecture des mesures d'un fichier AM}
\usage{
read.AM.mesures(file.AM, encoding = "macroman")
}
\arguments{
\item{encoding}{Pour les fichiers du magnétomètre, il faut "macroman" -> difficle à connaitre, peut être "latin1" ou "utf8".}
}
\description{
Lecture des mesures d'un fichier AM
}
