% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.AM.oldType.info}
\alias{read.AM.oldType.info}
\title{Lecture des mesures d'un fichier AM de l'ancien format}
\usage{
read.AM.oldType.info(file.AM, encoding = "macroman")
}
\value{
une data.frame avec les infos sur les spécimens
}
\description{
Lecture des mesures d'un fichier AM de l'ancien format
}
