% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.AM.oldType.mesures}
\alias{read.AM.oldType.mesures}
\title{Lecture des mesures d'un fichier AM de l'ancien format
La valeur de la température d'anisotropie n'était pas informée, il est donc possible de le renseigner avec ani.step.value}
\usage{
read.AM.oldType.mesures(file.AM, encoding = "macroman",
  ani.step.value = 700)
}
\arguments{
\item{ani.step.value}{valeur de la température des manips d'anisotropie}
}
\value{
une data.frame avec les infos sur les spécimens
}
\description{
Lecture des mesures d'un fichier AM de l'ancien format
La valeur de la température d'anisotropie n'était pas informée, il est donc possible de le renseigner avec ani.step.value
}
