% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.AM.orient}
\alias{read.AM.orient}
\title{Lecture des infos d'orientation sur mesures d'un fichier AM}
\usage{
read.AM.orient(file.AM, encoding = "macroman")
}
\value{
une data.frame avec les infos sur les spécimens
}
\description{
Lecture des infos d'orientation sur mesures d'un fichier AM
}
