% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.Pal.info}
\alias{read.Pal.info}
\title{Lecture des infos sur mesures d'un fichier AM}
\usage{
read.Pal.info(file.Pal, encoding = "macroman")
}
\value{
une data.frame avec les infos sur les spécimens
}
\description{
Lecture des infos sur mesures d'un fichier AM
}
