% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{read.Pal.mesures}
\alias{read.Pal.mesures}
\title{Lecture des mesures d'un fichier Pal (*.txt)}
\usage{
read.Pal.mesures(file.Pal, encoding = "macroman")
}
\arguments{
\item{encoding}{Pour les fichiers du magnétomètre, il faut "macroman" -> difficle à connaitre, peut être "latin1" ou "utf8".}
}
\value{
une data.frame
}
\description{
Lecture des mesures d'un fichier Pal (*.txt)
}
