% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{relocate.GAD}
\alias{relocate.GAD}
\title{Geocentric Axial Dipole
Reduction suivant le Dipôle Axiale Centré}
\usage{
relocate.GAD(I.site, lat.site, lat.reloc = "Paris")
}
\value{
I.reloc en degré
}
\description{
Geocentric Axial Dipole
Reduction suivant le Dipôle Axiale Centré
}
\examples{
relocate.GAD(65, 47.12)
}
\seealso{
\code{\link{relocate.VADM}}, \code{\link{relocate.VDM}}, \code{\link{relocate.VGP} }
}
