% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{relocate.VDM}
\alias{relocate.VDM}
\title{Virtual Dipôle Moment}
\usage{
relocate.VDM(
  F.site,
  I.site,
  D.site,
  lat.site,
  lon.site,
  lat.reloc = "Paris",
  lon.reloc = 0
)
}
\description{
Virtual Dipôle Moment
}
\examples{
relocate.VDM(55, 65,5, 47.12, 2.175,"Paris")
}
\seealso{
\code{\link{relocate.GAD}}, \code{\link{relocate.VADM}}, \code{\link{relocate.VGP} }
}
