% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{relocate.VGP}
\alias{relocate.VGP}
\title{Correction Virtual Geomagnetic Pole
Paris   lat:48.85 lon:2.3
Madrid  lat:40.4 lon:-3.68
Meriden lat:52.43 lon:-1.62
Athenes lat:37.96 lon:23.72}
\usage{
relocate.VGP(I.site, D.site, lat.site, lon.site, lat.reloc = "Paris",
  lon.reloc = 0)
}
\description{
Correction Virtual Geomagnetic Pole
Paris   lat:48.85 lon:2.3
Madrid  lat:40.4 lon:-3.68
Meriden lat:52.43 lon:-1.62
Athenes lat:37.96 lon:23.72
}
\examples{
relocate.VGP(65,5,DMS.to.DD(47,7,15),DMS.to.DD(2,10,12),"Paris")
}
\references{
Westphal, p.117 et Merill McElhinny, p.80
}
\seealso{
\code{\link{relocate.GAD, relocate.VADM, relocate.VDM }}
}
