% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{repliement.auto}
\alias{repliement.auto}
\title{repliement automatique'}
\usage{
repliement.auto(data, dec, aim = 1, inc.critique = 90)
}
\arguments{
\item{dec}{en degrés}

\item{inc}{en degrés}
}
\description{
repliement automatique'
}
