% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{repliement}
\alias{repliement}
\title{repliement
Calcul le repliement pour les trois position à plat "P", de chant "C" et debout "D"
pour les matériaux déplacés, la référence est le nord donc -angD}
\usage{
repliement(data, dec = NULL, aim = 1, name = NA, number = NA,
  position = "P")
}
\arguments{
\item{data}{valeur de l'inclinaison, ou une data.frame en degrés}

\item{dec}{en degrés}

\item{position}{trois valeurs posible à plat "P", de chant "C" et debout "D"}
}
\value{
en degrés
}
\description{
repliement
Calcul le repliement pour les trois position à plat "P", de chant "C" et debout "D"
pour les matériaux déplacés, la référence est le nord donc -angD
}
