% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{repliement.tranche}
\alias{repliement.tranche}
\title{repliement.tranche
Recherche la position qui permet d'avoir une position suivant la position debout ou dechant
donnant l'inclinaison la plus proche de la valeur inc.critique (en degrés)}
\usage{
repliement.tranche(
  data,
  dec = NULL,
  aim = 1,
  name = NULL,
  number = NULL,
  inc.critique = 90
)
}
\arguments{
\item{data}{valeur de l'inclinaison, ou une data.frame en degrés}

\item{dec}{en degrés}
}
\value{
I et D en degrés et X, Y, Z correspondant
}
\description{
repliement.tranche
Recherche la position qui permet d'avoir une position suivant la position debout ou dechant
donnant l'inclinaison la plus proche de la valeur inc.critique (en degrés)
}
