% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{rotation.mesure}
\alias{rotation.mesure}
\title{Fait tourner les mesures selon deviation}
\usage{
rotation.mesure(Data, deviation)
}
\arguments{
\item{Data}{un data.frame possédant la variable $step de type chr}

\item{deviation}{valeur de l'angle de rotation (en degrés).}
}
\description{
Fait tourner les mesures selon deviation
}
