% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{stat.mcFadden}
\alias{stat.mcFadden}
\title{Statistique de mcFadden sur l'inclinaison seul à partir des coordonées I et D}
\usage{
stat.mcFadden(inc, dec, inc.absolue = TRUE)
}
\arguments{
\item{inc}{liste des inclinaisons en degrés}

\item{inc.absolue}{calcul avec la valeur absolue des inclinaisons}

\item{des}{liste des déclinaisons en degrés}
}
\value{
en degrés
}
\description{
Statistique de mcFadden sur l'inclinaison seul à partir des coordonées I et D
}
