% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{sun.azimuth}
\alias{sun.azimuth}
\title{Calcul l'azimuth du soleil en un lieu à une date donnée à une heure "UTC"}
\usage{
sun.azimuth(
  Day,
  Month,
  Year,
  Hour,
  Minute,
  Seconde = 0,
  longdeg,
  longmin = 0,
  longsec = 0,
  latdeg,
  latmin = 0,
  latsec = 0
)
}
\description{
Calcul l'azimuth du soleil en un lieu à une date donnée à une heure "UTC"
}
\seealso{
\code{\link{https://codes-sources.commentcamarche.net/source/31774-calcul-de-la-position-du-soleil-declinaison-angle-horaire-altitude-et-azimut-altaz-solaire}}

\code{\link{https://fr.planetcalc.com/320/}}
}
