% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{supprime.etape}
\alias{supprime.etape}
\title{supprime les étapes à la valeur ani.etape.value et avec les sigles etape.sigle
Utiliser pour supprimer l'anisotropie}
\usage{
supprime.etape(Data, ani.etape.value = NULL, etape.sigle = c("Z+",
  "Z-", "X+", "X-", "Y+", "Y-", "ZB"))
}
\arguments{
\item{Data}{un data.frame possédant la variable $Etap de type chr}
}
\description{
supprime les étapes à la valeur ani.etape.value et avec les sigles etape.sigle
Utiliser pour supprimer l'anisotropie
}
