% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{to.cartesian.Z}
\alias{to.cartesian.Z}
\title{Valeur de Z pour I et D}
\usage{
to.cartesian.Z(inc, dec, aim = 1)
}
\arguments{
\item{inc}{liste des inclinaisons en degré}

\item{des}{liste des déclinaisons en degré}
}
\description{
Valeur de Z pour I et D
}
\seealso{
\code{\link{to.cartesian}} , \code{\link{to.polar}}
}
