% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{to.polar}
\alias{to.polar}
\title{Coordonnées polaires pour X, Y et Z}
\usage{
to.polar(X, Y, Z)
}
\value{
I, D, F en degré
}
\description{
Coordonnées polaires pour X, Y et Z
}
\seealso{
\code{\link{to.polar}} , \code{\link{cartesien}}
}
