% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{vecteur.partiel}
\alias{vecteur.partiel}
\title{Calcul les directions du vecteur partiel pour une série d'étape}
\usage{
vecteur.partiel(TabX, TabY, TabZ, en0 = TRUE)
}
\arguments{
\item{en0}{permet de calculer la composante qui passe par l'origine (0, 0)}
}
\value{
une data.frame "X", "Y", "Z", "I", "D", "F", "Sl", "MAD"
}
\description{
Calcul les directions du vecteur partiel pour une série d'étape
}
