% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{X}
\alias{X}
\title{Valeur de X graphique pour I et D
Fonction interne}
\usage{
X(i, d, ray = 1, i.min = 0, box.range)
}
\arguments{
\item{i}{inclinaison en degrés}

\item{d}{déclinaison en degrés}
}
\description{
Valeur de X graphique pour I et D
Fonction interne
}
