% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{zijderveld1}
\alias{zijderveld1}
\title{Trace un diagramme de Zijderveld type 1}
\usage{
zijderveld1(Data, Y = NULL, Z = NULL, pt.names = "", main = NULL,
  panel.first = NULL, pt.col = c("forestgreen", "blue3"),
  isometric = TRUE, ylim = NULL, legend.pos = NULL,
  legend.txt = c("(Y, X)", "(Y, Z)"), new = TRUE, ...)
}
\arguments{
\item{Data}{soit une data.frame avec les mesures X, Y et Z, soit que les valeurs de X}

\item{Y}{les valeurs de Y, si Data n''est pas une data.frame}

\item{Z}{les valeurs de Z, si Data n''est pas une data.frame}

\item{panel.first}{= grid() : affiche une grille}

\item{legend.pos}{= NULL : n'affiche pas la legende. legend.pos = "topleft" affiche en haut à gauche}

\item{pt.name}{= "": n'affiche rien. pt.name = NULL: affiche les étapes si Data est une data.frame}
}
\description{
Trace un diagramme de Zijderveld type 1
}
