% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{zijderveld1.T1T2}
\alias{zijderveld1.T1T2}
\title{trace un diagramme de zijderveld en calculant la composante entre les étapes T1 et T2}
\usage{
zijderveld1.T1T2(Data, T1 = NULL, T2 = NULL, withAni = FALSE,
  ani.etape.value = NULL, etape.sigle = c("Z+", "Z-", "X+", "X-", "Y+",
  "Y-", "ZB"), en0 = FALSE)
}
\description{
trace un diagramme de zijderveld en calculant la composante entre les étapes T1 et T2
}
