% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{zijderveld2}
\alias{zijderveld2}
\title{Trace un diagramme de Zijderveld type 2}
\usage{
zijderveld2(
  Data,
  Y = NULL,
  Z = NULL,
  pt.names = "",
  main = NULL,
  panel.first = NULL,
  pt.col = c("forestgreen", "blue3"),
  isometric = TRUE,
  ylim = NULL,
  legend.pos = NULL,
  new = TRUE,
  ...
)
}
\description{
Trace un diagramme de Zijderveld type 2
}
