% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{zjiderveld2}
\alias{zjiderveld2}
\title{Trace un diagramme de Zijderveld type 2}
\usage{
zjiderveld2(Data, Y = NULL, Z = NULL, pt.names = "", main = NULL,
  panel.first = NULL, pt.col = c("forestgreen", "blue3"),
  isometric = TRUE, ylim = NULL, legend.pos = NULL, new = TRUE,
  ...)
}
\description{
Trace un diagramme de Zijderveld type 2
}
