% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{angleD}
\alias{angleD}
\title{Calcul de l'angle D  entre  -pi <=  angD  < +3 pi  retourne en radian}
\usage{
angleD(X, Y)
}
\value{
angle en radian
Fonction interne
}
\description{
Calcul de l'angle D  entre  -pi <=  angD  < +3 pi  retourne en radian
}
