% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.eigen.tensors.all}
\alias{anisotropie.eigen.tensors.all}
\title{Calcul des tenseurs d'anisotropie pour tous les spécimens d'une liste de mesures}
\usage{
anisotropie.eigen.tensors.all(
  Data.mesures,
  Data.number,
  step.value,
  step.code = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB")
)
}
\value{
un data.frame avec les colonnes "L1", "L2", "L3", "L1.Inc", "L1.Dec", "L2.Inc", "L2.Dec", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul des tenseurs d'anisotropie pour tous les spécimens d'une liste de mesures
}
