% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{anisotropie.matrix.symetric}
\alias{anisotropie.matrix.symetric}
\title{Calcul la matrice d'anisotropie symetrisée et normalisé pour un spécimen
qui sert à la correction}
\usage{
anisotropie.matrix.symetric(mesures, step.value, step.code = c("Z+",
  "Z-", "X+", "X-", "Y+", "Y-", "ZB"), ...)
}
\arguments{
\item{mesures}{data.frame contenant les mesures}

\item{step.value}{typiquement la température des mesures d'anisotropie}

\item{step.code}{code indiquant les étapes. les noms peuvent changer, mais pas l'ordre}

\item{volume}{la valeur du volume du spécimen}

\item{TH}{la valeur du champ appliqué}
}
\value{
un data.frame avec les colonnes "L1", "L1.Inc", "L1.Dec", "L2", "L2.Inc", "L2.Dec", "L3", "L3.Inc", "L3.Dec", "F13", "F12", "F23"
}
\description{
Calcul la matrice d'anisotropie symetrisée et normalisé pour un spécimen
qui sert à la correction
}
