% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{arai}
\alias{arai}
\title{Standart calcul en plot to study paleo and archeo magnetic magnétisation
  by convention we use the two last character of the variable step to indicate the type of manip
  example 100RA , 100 is the temperature (step.value) , R is the sens of the magnetisation and A is the name of the step (step.name)}
\usage{
arai(
  mesures,
  relative = TRUE,
  verbose = TRUE,
  show.plot = TRUE,
  TH = 60,
  aim.coef = 1e-10 * 1e+06,
  step.J0 = "20N0",
  show.step.value = FALSE,
  R.mark = "R",
  V.mark = "V",
  P.mark = "P",
  L.mark = "L",
  Q.mark = "Q",
  pt.col = "blue",
  loop.col = "forestgreen",
  begin.step.value = 0,
  end.step.value = 1000
)
}
\description{
Standart calcul en plot to study paleo and archeo magnetic magnétisation
  by convention we use the two last character of the variable step to indicate the type of manip
  example 100RA , 100 is the temperature (step.value) , R is the sens of the magnetisation and A is the name of the step (step.name)
}
\references{
Coe 1978 : DOI: 10.1029/JB083iB04p01740
Prévost et Al. 1985 DOI: 10.1029/JB090iB12p10417
}
