% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{DMS.to.DD}
\alias{DMS.to.DD}
\title{Conversion de Degrés Minute Second en Degré Décimal}
\usage{
DMS.to.DD(degre, minute, second = 0)
}
\arguments{
\item{degre}{degrés entier}

\item{minute}{minute entière}

\item{second}{seconde en décimal}
}
\description{
Conversion de Degrés Minute Second en Degré Décimal
}
\seealso{
\code{\link{DD.to.DMS}}
}
