% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{find.extremum}
\alias{find.extremum}
\title{Valeurs maximun d'un tracer lambert
Permet de trouver les valeurs minimales et maximales pour tracer un lambert
Fonction interne}
\usage{
find.extremum(i.min = 0, i.max = 90, d.min = -90, d.max = 270)
}
\description{
Valeurs maximun d'un tracer lambert
Permet de trouver les valeurs minimales et maximales pour tracer un lambert
Fonction interne
}
