% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{julian.day}
\alias{julian.day}
\title{Le nombre de jour julien pour les calculs astronomiques
Le jour julien 0 commence le 24 novembre -4713 (4712 BC) à 12h}
\usage{
\method{julian}{day}(jour, mois, annee, heure, minute, seconde)
}
\description{
Le nombre de jour julien pour les calculs astronomiques
Le jour julien 0 commence le 24 novembre -4713 (4712 BC) à 12h
}
\seealso{
\code{\link{https://codes-sources.commentcamarche.net/source/31774-calcul-de-la-position-du-soleil-declinaison-angle-horaire-altitude-et-azimut-altaz-solaire}}
}
