% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{lambert.ID.tensors}
\alias{lambert.ID.tensors}
\title{Tracer dans le diagramme Lambert/ Stéréo d'un tenseur}
\usage{
lambert.ID.tensors(Data, point.col = "blue3", new = TRUE, ...)
}
\arguments{
\item{Data}{un data.frame avec les variables "L1.Inc", "L1.Dec", "L2.Inc", "L2.Dec", "L3.Inc", "L3.Dec"}
}
\description{
Tracer dans le diagramme Lambert/ Stéréo d'un tenseur
}
