% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{plot.desaim}
\alias{plot.desaim}
\title{Tracer d'un diagramme de désaimantation}
\usage{
\method{plot}{desaim}(Data, F = NULL, point.col = "blue3", pch = 21,
  type = "b", xlab = "°C", ylab = "", main = NULL, names = NA,
  normalize = TRUE, new = TRUE, ...)
}
\arguments{
\item{normalize}{permet de comparer l'évolution quelque soit l'amplitude}
}
\description{
Tracer d'un diagramme de désaimantation
}
