% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{polaire}
\alias{polaire}
\title{Coordonnées polaires pour X, Y et Z}
\usage{
polaire(X, Y, Z)
}
\value{
I, D, F en degré
}
\description{
Coordonnées polaires pour X, Y et Z
}
\seealso{
\code{\link{calcul.vecteur.polaire}} , \code{\link{cartesien}}
}
