% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{relocate.GAD}
\alias{relocate.GAD}
\title{Geocentric Axial Dipole
Reduction suivant le Dipôle Axiale Centré}
\usage{
relocate.GAD(I.site, lat.site, lat.reloc = "Paris")
}
\arguments{
\item{lat.reloc}{Latitude de réduction. On peut aussi utiliser "Paris" pour 48.85
"Madrid" pour 40.4; "Meriden" pour 52.43; "Athenes" pour 37.96}
}
\value{
I.reloc en degré
}
\description{
Geocentric Axial Dipole
Reduction suivant le Dipôle Axiale Centré
}
\examples{
relocate.GAD(65, 47.12)
}
\seealso{
\code{\link{relocate.VADM}}, \code{\link{relocate.VDM}}, \code{\link{relocate.VGP} }
}
