% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{relocate.VADM}
\alias{relocate.VADM}
\title{Virtual Axial Dipole Moment VADM}
\usage{
relocate.VADM(F.site, lat.site, lat.reloc = "Paris")
}
\description{
Virtual Axial Dipole Moment VADM
}
\examples{
Reduction par defaut à Paris
 relocate.VADM(55, 47.12)
 relocate.VADM(55, 47.12, "Paris")
 relocate.VADM(55, 47.12, lat.reloc = "48.85")
}
\seealso{
\code{\link{relocate.GAD}}, \code{\link{relocate.VDM}}, \code{\link{relocate.VGP} }
}
