% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{remove.step}
\alias{remove.step}
\title{Supprime les étapes à la valeur ani.step.value et avec les codes step.code
Utiliser par défaut pour remover les étapes d'anisotropie}
\usage{
remove.step(Data, ani.step.value = NULL, step.code = c("Z+", "Z-",
  "X+", "X-", "Y+", "Y-", "ZB"))
}
\arguments{
\item{Data}{un data.frame possédant la variable $step de type chr}

\item{ani.step.value}{valeur de la température d'anisotropie. Par defaut NUll, alors la tempértature est retrouvée automatiquement avec les step.code définis}

\item{step.code}{chaîne de caractère représentant les étapes de l'anisotropie "Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB"}
}
\description{
Supprime les étapes à la valeur ani.step.value et avec les codes step.code
Utiliser par défaut pour remover les étapes d'anisotropie
}
