% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{remove.step}
\alias{remove.step}
\title{Supprime les étapes à la valeur step.value et avec les codes step.code
Utiliser par défaut pour remover les étapes d'anisotropie}
\usage{
remove.step(Data, step.value = NULL, step.code = c("Z+", "Z-", "X+",
  "X-", "Y+", "Y-", "ZB"))
}
\arguments{
\item{Data}{un data.frame possédant la variable $step de type chr}

\item{step.value}{valeur de la température . Par defaut NUll, alors la tempértature est retrouvée automatiquement avec les step.code définis. Si step.value = '', alors suppression de toutes les étapes avec le step.code}

\item{step.code}{chaîne de caractère représentant par exemple les étapes de l'anisotropie "Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB", ou des erreurs "??"}
}
\description{
Supprime les étapes à la valeur step.value et avec les codes step.code
Utiliser par défaut pour remover les étapes d'anisotropie
}
