% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{stat.fisher}
\alias{stat.fisher}
\title{Statistique de Fisher
modifié, pas de pondération calcul de A95 Vrai sans simplification tel que fisher 1953}
\usage{
stat.fisher(Data, dec = NULL, aim = NA, pfish = 0.95, inc.absolue = TRUE)
}
\arguments{
\item{Data}{liste des inclinaisons en degré ou une data.frame avec les variables $I et $D}

\item{dec}{liste des déclinaisons en degrés}

\item{aim}{liste des aimantations, facultatif}

\item{pfish}{pourcentage de confiance}

\item{inc.absolue}{calcul avec la valeur absolue des inclinaisons}
}
\value{
en degrés
}
\description{
Statistique de Fisher
modifié, pas de pondération calcul de A95 Vrai sans simplification tel que fisher 1953
}
\seealso{
\code{\link{stat.mcFadden}}
}
\keyword{fisher}
