% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{stat.mcFadden}
\alias{stat.mcFadden}
\title{Statistique de mcFadden sur l'inclinaison seul à partir des coordonées I et D}
\usage{
stat.mcFadden(inc, dec, inc.absolue = TRUE, digits = 3)
}
\description{
Statistique de mcFadden sur l'inclinaison seul à partir des coordonées I et D
}
