% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{sun.azimuth}
\alias{sun.azimuth}
\title{Calculate the azimuth of the sun in a place at a given date at a given time "UTC".}
\usage{
sun.azimuth(
  day,
  month,
  year,
  hour,
  minute,
  seconde = 0,
  longdeg,
  longmin = 0,
  longsec = 0,
  latdeg,
  latmin = 0,
  latsec = 0
)
}
\description{
Calculate the azimuth of the sun in a place at a given date at a given time "UTC".
}
\seealso{
\code{\link{https://codes-sources.commentcamarche.net/source/31774-calcul-de-la-position-du-soleil-declinaison-angle-horaire-altitude-et-azimut-altaz-solaire}}

\code{\link{https://fr.planetcalc.com/320/}}
}
