% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArMag.R
\name{zijderveld2.T1T2}
\alias{zijderveld2.T1T2}
\title{trace un diagramme de zijderveld en calculant la composante entre les étapes T1 et T2}
\usage{
zijderveld2.T1T2(Data, T1 = NULL, T2 = NULL, show.step = FALSE,
  ignore.ani = TRUE, ani.step.value = NULL, main = "",
  step.code = c("Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB"),
  legend.pos = NULL, legend.txt = c("(Y, X)", "(Y, Z)"), en0 = FALSE)
}
\arguments{
\item{T1}{correspond à step.value ou température la plus basse}

\item{T2}{correspond à step.value ou température la plus haute}

\item{show.step}{booléen permettant l'affichage des étapes}

\item{ani.step.value}{correspond à l'step.value ou la température de la détermination de l'anisotropie}

\item{step.code}{chaîne de caractère représentant les étapes de l'anisotropie "Z+", "Z-", "X+", "X-", "Y+", "Y-", "ZB". Cette ordre est obligatoire}

\item{en0}{booléen permettant de forcer la composante partielle de passer par l'origine}

\item{TRUE}{permet de voir les étapes d'anisotropie}
}
\description{
trace un diagramme de zijderveld en calculant la composante entre les étapes T1 et T2
}
\seealso{
\code{\link{composante.partielle.T1T2, zijderveld2.T1T2}}
}
